🔧 Rename app entrypoint
